changed dtype to long

+++ agent/flow_match_estimation.py
@@ -23,7 +23,7 @@ def estimate_fm_value(states, imged_rewards):
     estimator = LogEdgeFlowEstimator(env=env, module_name="NeuralNet")
     parametrization = FMParametrization(logF=estimator)
     loss_fn = FlowMatching(parametrization=parametrization)
-    states_container = env.make_States_class()(states_tensor=states)
+    states_container = env.make_States_class()(states_tensor=states.to(torch.int64))
     return loss_fn(states_container, imged_rewards)
 
 class FlowMatching(StateDecomposableLoss):
